email-marketing/requirements.txt
Django==2.2.7
django-filter==2.2.0
djangorestframework==3.10.3
Markdown==3.1.1
pytz==2019.3
sqlparse==0.3.0
psycopg2==2.8.4
requests==2.22.0
celery==4.3.0
django-rest-swagger==2.2.0